# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evaluaciones OMR API",
    "description": "Servicio de evaluaciones con reconocimiento óptico de marcas (OMR).\n\n**Flujo principal**\n1. Crear una evaluación (banco de preguntas) → agregar secciones → cargar preguntas con `seccion_id`.\n2. Generar hojas: `POST /generar/{evaluacion_id}` con una selección **template** (reparto por\n   cantidad/porcentaje por sección) o **estructurada** (banco completo o subconjunto ordenado).\n   El `identificador` es opcional: si falta se autogenera (`{short_id}-{n}`) y la reimpresión\n   con el mismo identificador devuelve la misma variante.\n3. Escanear la hoja llena y evaluarla con `POST /evaluar` (califica contra el banco).\n\n**Hoja genérica (sin examen)**\n- `POST /generar/generica` genera solo burbujas (N preguntas × M opciones, 2–5).\n- `POST /leer` devuelve lo llenado (identificador persona + burbujas marcadas) sin contrastar nada.\n\nVer **Redoc** (`/redoc`) para la referencia completa o el esquema OpenAPI en `/openapi.json`.",
    "contact": {
      "name": "Evaluaciones OMR"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.1.0"
  },
  "paths": {
    "/api/v1/evaluacion": {
      "post": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Crear Evaluacion",
        "description": "Crea una evaluación (banco de preguntas). Genera un `short_id` único usado\ncomo prefijo del identificador autogenerado de las hojas (D4).",
        "operationId": "crear_evaluacion_api_v1_evaluacion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluacionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evaluacion/{evaluacion_id}": {
      "get": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Obtener Evaluacion",
        "description": "Obtiene una evaluación por su id.",
        "operationId": "obtener_evaluacion_api_v1_evaluacion__evaluacion_id__get",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Eliminar Evaluacion",
        "description": "Elimina una evaluación y todo lo asociado (preguntas, secciones, hojas, resultados).",
        "operationId": "eliminar_evaluacion_api_v1_evaluacion__evaluacion_id__delete",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluacion/{evaluacion_id}/secciones": {
      "post": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Crear Secciones Batch",
        "description": "Crea varias secciones en el banco de una evaluación (caso 1: reparto por secciones).",
        "operationId": "crear_secciones_batch_api_v1_evaluacion__evaluacion_id__secciones_post",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeccionBatchCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeccionResponse"
                  },
                  "title": "Response Crear Secciones Batch Api V1 Evaluacion  Evaluacion Id  Secciones Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Listar Secciones",
        "description": "Lista las secciones de un banco, ordenadas por su `orden`.",
        "operationId": "listar_secciones_api_v1_evaluacion__evaluacion_id__secciones_get",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeccionResponse"
                  },
                  "title": "Response Listar Secciones Api V1 Evaluacion  Evaluacion Id  Secciones Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seccion/{seccion_id}": {
      "get": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Obtener Seccion",
        "description": "Obtiene una sección del banco.",
        "operationId": "obtener_seccion_api_v1_seccion__seccion_id__get",
        "parameters": [
          {
            "name": "seccion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Seccion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeccionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Actualizar Seccion",
        "description": "Actualiza nombre/descripción/orden de una sección.",
        "operationId": "actualizar_seccion_api_v1_seccion__seccion_id__put",
        "parameters": [
          {
            "name": "seccion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Seccion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeccionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeccionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Eliminar Seccion",
        "description": "Elimina una sección y sus preguntas (cascade).",
        "operationId": "eliminar_seccion_api_v1_seccion__seccion_id__delete",
        "parameters": [
          {
            "name": "seccion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Seccion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluacion/{evaluacion_id}/hojas": {
      "get": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Listar Hojas",
        "description": "Lista las hojas de respuestas generadas para una evaluación (recientes primero).",
        "operationId": "listar_hojas_api_v1_evaluacion__evaluacion_id__hojas_get",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HojaRespuestaInfo"
                  },
                  "title": "Response Listar Hojas Api V1 Evaluacion  Evaluacion Id  Hojas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluacion/{evaluacion_id}/preguntas/batch": {
      "post": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Crear Preguntas Batch",
        "description": "Carga preguntas en el banco. Cada pregunta debe tener entre 2 y 5 opciones (D6)\ny puede asociarse a una sección con `seccion_id` (o texto legacy `seccion`).",
        "operationId": "crear_preguntas_batch_api_v1_evaluacion__evaluacion_id__preguntas_batch_post",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreguntaBatchCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreguntaResponse"
                  },
                  "title": "Response Crear Preguntas Batch Api V1 Evaluacion  Evaluacion Id  Preguntas Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluacion/{evaluacion_id}/preguntas": {
      "get": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Listar Preguntas",
        "description": "Lista las preguntas de un banco, ordenadas por (sección, orden).",
        "operationId": "listar_preguntas_api_v1_evaluacion__evaluacion_id__preguntas_get",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreguntaResponse"
                  },
                  "title": "Response Listar Preguntas Api V1 Evaluacion  Evaluacion Id  Preguntas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pregunta/{pregunta_id}": {
      "delete": {
        "tags": [
          "Banco de preguntas"
        ],
        "summary": "Eliminar Pregunta",
        "description": "Elimina una pregunta del banco junto con sus opciones.",
        "operationId": "eliminar_pregunta_api_v1_pregunta__pregunta_id__delete",
        "parameters": [
          {
            "name": "pregunta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pregunta Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EvaluacionCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "nombre"
        ],
        "title": "EvaluacionCreate"
      },
      "EvaluacionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "short_id": {
            "type": "string",
            "title": "Short Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "short_id",
          "nombre",
          "descripcion"
        ],
        "title": "EvaluacionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HojaRespuestaInfo": {
        "properties": {
          "hoja_id": {
            "type": "string",
            "format": "uuid",
            "title": "Hoja Id"
          },
          "evaluacion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evaluacion Id"
          },
          "identificador": {
            "type": "string",
            "title": "Identificador"
          },
          "qr_data": {
            "type": "object",
            "title": "Qr Data"
          },
          "pdf_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Path"
          },
          "cantidad_preguntas": {
            "type": "integer",
            "title": "Cantidad Preguntas",
            "default": 0
          },
          "config_seleccion": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Seleccion"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "hoja_id",
          "evaluacion_id",
          "identificador",
          "qr_data",
          "pdf_path",
          "created_at"
        ],
        "title": "HojaRespuestaInfo"
      },
...
      "PreguntaBatchCreate": {
        "properties": {
          "preguntas": {
            "items": {
              "$ref": "#/components/schemas/PreguntaCreate"
            },
            "type": "array",
            "title": "Preguntas"
          }
        },
        "type": "object",
        "required": [
          "preguntas"
        ],
        "title": "PreguntaBatchCreate"
      },
...
      "SeccionBatchCreate": {
        "properties": {
          "secciones": {
            "items": {
              "$ref": "#/components/schemas/SeccionCreate"
            },
            "type": "array",
            "title": "Secciones",
            "description": "Lista de secciones a crear"
          }
        },
        "type": "object",
        "required": [
          "secciones"
        ],
        "title": "SeccionBatchCreate"
      },
      "SeccionCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre",
            "description": "Nombre de la sección (p. ej. 'Álgebra')"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion",
            "description": "Descripción libre de la sección"
          },
          "orden": {
            "type": "integer",
            "title": "Orden",
            "description": "Posición en que se imprime la sección dentro del banco",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "nombre"
        ],
        "title": "SeccionCreate"
      },
...
      "SeccionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "orden": {
            "type": "integer",
            "title": "Orden"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "descripcion",
          "orden"
        ],
        "title": "SeccionResponse"
      },
      "SeccionUpdate": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre",
            "description": "Nuevo nombre de la sección"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion",
            "description": "Nueva descripción"
          },
          "orden": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orden",
            "description": "Nueva posición (secciones mayores se desplazan)"
          }
        },
        "type": "object",
        "title": "SeccionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}